% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_chat.R
\name{peer_connect}
\alias{peer_connect}
\title{Connect to the PeerChat server}
\usage{
peer_connect(username, password)
}
\arguments{
\item{username}{Character string. The username to register with on the server.}

\item{password}{Character string. The password that other users needs to know to message you.}
}
\value{
Invisible \code{NULL}. Raises an error on failure.
}
\description{
Establishes a UDP connection to the synchronization server and registers
the client with the given username and password. Must be called before
any other PeerChat function. It also creates hashmap that stores all necessary info.
}
\examples{
\dontrun{
peer_connect("nizwan", "password123")
}
}
