% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_chat.R
\name{peer_get_fd}
\alias{peer_get_fd}
\title{Get the socket file descriptor}
\usage{
peer_get_fd()
}
\value{
Integer. The socket file descriptor, or \code{-1} if not connected.
}
\description{
Returns the internal UDP socket file descriptor. Intended for advanced
use cases such as custom select loops.
}
\examples{
\dontrun{
fd <- peer_get_fd()
}
}
