% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_chat.R
\name{peer_get_user}
\alias{peer_get_user}
\title{Discover a peer by username}
\usage{
peer_get_user(username, password)
}
\arguments{
\item{username}{Character string. The username of the peer to discover.}

\item{password}{Character string. The password of the peer.}
}
\value{
Invisible \code{NULL}. Raises an error on failure.
}
\description{
Sends a request to the server to look up a peer's network address.
On success, the peer is stored internally and later on can be messaged or pinged.
Requires prior "registration" via \code{\link{peer_connect}}.
}
\examples{
\dontrun{
peer_get_user("max", "max_password")
}
}
