% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeerChat-package.R
\docType{package}
\name{PeerChat-package}
\alias{PeerChat}
\alias{PeerChat-package}
\title{PeerChat: Peer-to-Peer Messaging Library}
\description{
UDP peer-to-peer messaging library backed by a C networking core. Provides R bindings for creating sockets, discovering peers via a central synchronization server written in C, and exchanging messages directly between clients without relaying through the server. It leverages a custom made data protocol.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nizwant/computational_data_science}
  \item \url{https://nizwant.github.io/computational_data_science/}
  \item Report bugs at \url{https://github.com/nizwant/computational_data_science/issues}
}

}
\author{
\strong{Maintainer}: Mateusz Nizwantowski \email{mateusznizwantowski@gmail.com}

Authors:
\itemize{
  \item Mateusz Nizwantowski \email{mateusznizwantowski@gmail.com}
}

}
\keyword{internal}
