% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_chat.R
\name{peer_send_ping}
\alias{peer_send_ping}
\title{Ping a peer}
\usage{
peer_send_ping(recipient)
}
\arguments{
\item{recipient}{Character string. The username of the peer to ping.}
}
\value{
Invisible \code{NULL}. Raises an error if the peer is not found
  or the send fails.
}
\description{
Sends a ping packet to a previously discovered peer, works also on server.
}
\examples{
\dontrun{
peer_send_ping("max_my_friend")
}
}
